pyproject.toml
[tool.poetry]
name = "metis-rag"
version = "0.1.0"
description = ""
authors = ["Adegbenga Agoro <adegbenga@designsbycrenet.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.37.1"
boto3 = "^1.34.131"
python-dotenv = "^1.0.1"
botocore = "^1.35.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
